% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-exposure.r
\name{apply_exposure_concurrent}
\alias{apply_exposure_concurrent}
\title{apply exposure for concurrent}
\usage{
apply_exposure_concurrent(x, unit_var, time_var, treated_units)
}
\description{
apply exposure for concurrent
}
