% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-exposure.r
\name{apply_exposure_single}
\alias{apply_exposure_single}
\title{apply exposure not concurrent}
\usage{
apply_exposure_single(
  x,
  unit_var,
  time_var,
  treated_units,
  change_code_treatment
)
}
\description{
apply exposure not concurrent
}
