% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-treatment-effect.r
\name{apply_treatment_effect.linear}
\alias{apply_treatment_effect.linear}
\title{apply treatment effect for linear model}
\usage{
\method{apply_treatment_effect}{linear}(ConfigObject, te)
}
\description{
apply treatment effect for linear model
}
