% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-treatment-effect.r
\name{apply_treatment_effect.log}
\alias{apply_treatment_effect.log}
\title{apply treatment effect for log model}
\usage{
\method{apply_treatment_effect}{log}(ConfigObject, te)
}
\description{
apply treatment effect for log model
}
