% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-treatment-effect.r
\name{apply_treatment_effect.loglinear}
\alias{apply_treatment_effect.loglinear}
\title{apply treatment effect for log-linear model}
\usage{
\method{apply_treatment_effect}{loglinear}(ConfigObject, te)
}
\description{
apply treatment effect for log-linear model
}
