% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-treatment-effect-concurrent.R
\name{apply_treatment_effect_concurrent.linear}
\alias{apply_treatment_effect_concurrent.linear}
\title{apply treatment effect for linear model}
\usage{
\method{apply_treatment_effect_concurrent}{linear}(ConfigObject, te)
}
\description{
apply treatment effect for linear model
}
