% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-treatment-effect-concurrent.R
\name{apply_treatment_effect_concurrent.log}
\alias{apply_treatment_effect_concurrent.log}
\title{apply treatment effect for log model}
\usage{
\method{apply_treatment_effect_concurrent}{log}(ConfigObject, te)
}
\description{
apply treatment effect for log model
}
