% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-treatment-effect-concurrent.R
\name{apply_treatment_effect_concurrent.loglinear}
\alias{apply_treatment_effect_concurrent.loglinear}
\title{apply treatment effect for log-linear model}
\usage{
\method{apply_treatment_effect_concurrent}{loglinear}(ConfigObject, te)
}
\description{
apply treatment effect for log-linear model
}
