% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-treatment-effect-concurrent.R
\name{apply_treatment_effect_concurrent}
\alias{apply_treatment_effect_concurrent}
\title{apply specified treatment effect and transform outcome to match model type}
\usage{
apply_treatment_effect_concurrent(ConfigObject, te)
}
\arguments{
\item{ConfigObject}{R6Class object of class "OpticConfig"}

\item{te}{output of effect_magnitude, the true effect}
}
\description{
Calls appropriate method depending on model type (class) of OpticConfig object
}
