% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-constant.r
\name{calc_constant}
\alias{calc_constant}
\title{constant needed for slow coding}
\usage{
calc_constant(m)
}
\arguments{
\item{m}{month integer}
}
\description{
sum of vector 1:m, when m (month) is not zero
}
