% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-exposure.r
\name{calculate_exposure}
\alias{calculate_exposure}
\title{Calculates the exposure rate applied to each year provided month
of policy implementation and number of years to full implementation}
\usage{
calculate_exposure(month, n_years, monthly_effect = (1/n_years)/12)
}
\arguments{
\item{month}{month of year (as integer) that policy takes effect}

\item{n_years}{number of months until full implementation in effect}

\item{month_effect}{increment of exposure to apply each month; default is
((1/n_years) / 12) (constant over the period)}
}
\description{
Calculates the exposure rate applied to each year provided month
of policy implementation and number of years to full implementation
}
