% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concurrent-methods.R
\name{concurrent_model}
\alias{concurrent_model}
\title{run model and store results}
\usage{
concurrent_model(model_simulation)
}
\description{
runs the model against the prepared data along with
    any provided arguments. Stores the model object in the input
    list, new element named "model_result" and returns full list
}
