% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concurrent-methods.R
\name{concurrent_postmodel}
\alias{concurrent_postmodel}
\title{process results from model(s)}
\usage{
concurrent_postmodel(model_simulation)
}
\description{
summarizes the statistical performance of the model(s) being compared by computing summary information on the model fit, estimated effects and standard errors
}
