% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concurrent-methods.R
\name{concurrent_premodel}
\alias{concurrent_premodel}
\title{pre-modeling method to apply to config object/data}
\usage{
concurrent_premodel(model_simulation)
}
\description{
since there are different data transformations needed for different
    model approaches (e.g., linear, count, log-linear), the treatment effect application and other data prep steps
    are run here before modeling in the model-specific config object rather 
    than in the sampling step
}
