% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concurrent-methods.R
\name{concurrent_results}
\alias{concurrent_results}
\title{compiles the final results}
\usage{
concurrent_results(r)
}
\description{
compiles the results into one table for all permutations of the simulation
}
