% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concurrent-methods.R
\name{concurrent_sample}
\alias{concurrent_sample}
\title{perform sampling and coding of treatment for concurrent policy simulations}
\usage{
concurrent_sample(single_simulation)
}
\arguments{
\item{single_simulation}{object created from SimConfig$setup_single_simulation()}
}
\description{
todo
}
