% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-config.R
\name{configure_simulation}
\alias{configure_simulation}
\title{Create a configuration object used to run simulations}
\usage{
configure_simulation(
  x,
  unit_var,
  time_var,
  model_call,
  model_formula,
  effect_magnitude,
  n_units,
  iters = 5000,
  effect_direction = c("null", "pos", "neg"),
  model_args = NULL,
  policy_speed = c("instant"),
  n_implementation_periods = NA,
  set_seed = FALSE,
  time_period_restriction = NULL,
  lag_outcome = FALSE,
  se_adjust = c("cluster", "huber", "huber-cluster"),
  concurrent = FALSE,
  correction_factors = NULL,
  rhos = NULL,
  change_code_treatment = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{x}{data.frame to use for model simulation}

\item{unit_var}{variable in x that represents unit of observation (e.g., state, person, record), character name}

\item{time_var}{variable in x that represents time (e.g., year, month), character name}

\item{model_call}{currently supported options are "lm", "glm", "glm.nb"}

\item{model_formula}{formula or list of formulas to iterate for simulation}

\item{effect_magnitude}{use for non-null effect directions to modify treatment outcome}

\item{n_units}{number of units to apply treatment to; may provide vector of integers for multiple sims}

\item{iters}{number of iterations for each sim, default is 5,000}

\item{effect_direction}{one or more of "null", "pos", "neg"; if "pos" or "neg" are supplied and no correction
factors are also provided then the simulation will throw a warning and run the null model first anyway to
get correction factors}

\item{model_args}{optional list of additional arguments to pass to model_call}

\item{policy_speed}{one or more of "instant", "slow"; if "slow" then must also provide value for
n_implementation_periods}

\item{n_implementation_periods}{number of time periods (time_var) to phase in treatment from 0 to 1
for "slow" policy speed}

\item{set_seed}{if TRUE will use seed and seed will be included in results for reproducibility;
default is FALSE}

\item{time_period_restriction}{provide vector of values that exist in time_var to restrict treatmenent
so it only can occur in these time periods}

\item{lag_outcome}{whether or not to include a lag (1 time period lag) of the outcome variable in
the model}

\item{se_adjust}{which standard error adjustments to produce, one or more of "cluster", "huber",
"huber-cluster"}

\item{concurrent}{TRUE if this be a simulation that tests concurrent treatment policies, default
is FALSE}

\item{correction_factors}{provide correction factors for "pos", "neg" effect directions if not
running "null"; if running "null" this will be ignored and if needed and not provided
will run "null" even if not specified in effect_direction}

\item{rhos}{requires if concurrent is TRUE, different correlations between start dates of
policies to iterate over}

\item{change_code_treatment}{boolean for whether or not to code treatment as a change in treatment
from prior period (used for autoregressive modeling), default is FALSE}

\item{verbose}{should I be chatty? Default is yes, I am chatty.}
}
\description{
Performs validation on inputs and produces a configuration object
    that contains all required parameters to dispatch simulation runs for data
    provided.
}
