% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct-rejection-rate.r
\name{correct_rejection_rate}
\alias{correct_rejection_rate}
\title{formula for correcting p-values using correction factor}
\usage{
correct_rejection_rate(coeffs, ses, cf, effect_direction = "null")
}
\arguments{
\item{coeffs}{vector of regression coefficients}

\item{ses}{vector of standard errors related to provided coefficients}

\item{cf}{correction factor to use for adjustment}

\item{effect_direction}{direction of true effect, one of "null", "neg", "pos"}
}
\description{
formula for correcting p-values using correction factor
}
