% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correction-factor.r
\name{correction_factor}
\alias{correction_factor}
\title{calculate the correction factor}
\usage{
correction_factor(test_stats, type = "t")
}
\arguments{
\item{test_stats}{test statistics}

\item{type}{what type of test statistic is being supplied; one of "t", "wald"; default is "t"}
}
\description{
calculate the correction factor
}
