% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch-simulations.R
\name{dispatch_simulations}
\alias{dispatch_simulations}
\title{Dispatch Simulations}
\usage{
dispatch_simulations(
  object,
  seed = NULL,
  use_future = FALSE,
  verbose = 0,
  graceful = FALSE,
  ...
)
}
\arguments{
\item{object}{Simulation scenarios object created using `optic_simulation`}

\item{seed}{Specified as either NULL or a numeric. Sets a seed, which is becomes an index in results, for each independent set of simulations in `optic_simulation`.}

\item{use_future}{Runs simulation scenarios in parallel. Default FALSE, set to TRUE if you would like for the iterations to be run in parallel.}

\item{verbose}{Default 0. If greater than 0, provides details on what is currently running.}

\item{graceful}{If TRUE, errors in iterations are caught and retried up to 10 percent of total iterations. If FALSE, errors are not caught and will stop execution. Default is FALSE.}

\item{...}{Additional parameters to be passed to model call.}
}
\value{
A single `data.frame` containing estimated treatment effects and summary statistics by model and draw, bound together to allow for different column names.
}
\description{
Runs optic simulations
}
