% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect-magnitudes.r
\name{effect_magnitude.linear}
\alias{effect_magnitude.linear}
\title{calculate true effect for linear model}
\usage{
\method{effect_magnitude}{linear}(ConfigObject)
}
\description{
calculate true effect for linear model
}
