% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect-magnitudes.r
\name{effect_magnitude.log}
\alias{effect_magnitude.log}
\title{calculate true effect for log model}
\usage{
\method{effect_magnitude}{log}(ConfigObject)
}
\description{
calculate true effect for log model
}
