% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect-magnitudes.r
\name{effect_magnitude.loglinear}
\alias{effect_magnitude.loglinear}
\title{calculate true effect for log-linear model}
\usage{
\method{effect_magnitude}{loglinear}(ConfigObject)
}
\description{
calculate true effect for log-linear model
}
