% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect-magnitudes.r
\name{effect_magnitude}
\alias{effect_magnitude}
\title{calculate the true effect for an OpticConfig object}
\usage{
effect_magnitude(ConfigObject)
}
\arguments{
\item{ConfigObject}{R6Class object of class OpticConfig}
}
\description{
calculate the true effect for an OpticConfig object
}
