% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-list.R
\name{exposure_list}
\alias{exposure_list}
\title{create exposure values across time periods for sampled treated units}
\usage{
exposure_list(
  sampled_time_period,
  mo,
  available_periods,
  policy_speed,
  n_implementation_periods
)
}
\arguments{
\item{sampled_time_period}{year that policy is first enacted}

\item{mo}{month that policy is first enacted}
}
\description{
create exposure values across time periods for sampled treated units
}
