% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-effect-magnitudes.r
\name{generate_effect_magnitudes}
\alias{generate_effect_magnitudes}
\title{calculate the effect magnitude for varying model types}
\usage{
generate_effect_magnitudes(target_deaths, total_population, number_years,
  model_type)
}
\arguments{
\item{model_type}{"linear", "log", or "log-lin"}
}
\description{
calculate the effect magnitude for varying model types
}
