% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-concurrent-distance.R
\name{get_distance_avgs}
\alias{get_distance_avgs}
\title{get distance between concurrent policies for each treated unit and summarize}
\usage{
get_distance_avgs(treated_units, units = "days")
}
\description{
get distance between concurrent policies for each treated unit and summarize
}
