% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-treated-units.r
\name{get_treated_units}
\alias{get_treated_units}
\title{get treated units}
\usage{
get_treated_units(
  x,
  n,
  unit_var,
  time_var,
  policy_speed,
  n_implementation_periods,
  concurrent,
  rho,
  time_period_restriction
)
}
\description{
get treated units
}
