% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iter-results.r
\name{iter_results.lm}
\alias{iter_results.lm}
\title{TODO: write documentation for this function if the user needs to see this.}
\usage{
iter_results.lm(model_simulation)
}
\arguments{
\item{model_simulation}{}
}
\description{
TODO: write documentation for this function if the user needs to see this.
}
