% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iter-results.r
\name{iter_results}
\alias{iter_results}
\title{iter results}
\usage{
iter_results(model_simulation)
}
\arguments{
\item{model_simulation}{R6 class object for simulation config}

\item{m}{model object}
}
\description{
iter results
}
