% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-terms.r
\name{model_terms}
\alias{model_terms}
\title{parse a formula object into its left-hand-side and right-hand-side components}
\usage{
model_terms(x)
}
\arguments{
\item{x}{formula to parse}
}
\value{
list with named elements "lhs" and "rhs"
}
\description{
parse a formula object into its left-hand-side and right-hand-side components
}
