% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no-confounding-methods.R
\name{noconf_premodel}
\alias{noconf_premodel}
\title{pre-modeling method to apply to config object/data}
\usage{
noconf_premodel(model_simulation)
}
\description{
since there are different data transformations needed for different
    model approaches, the treatment effect application and other data prep steps
    are run here before modeling but in the model-specific config object rather 
    than in the sampling step that would apply to all models in the sim
}
