% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no-confounding-methods.R
\name{noconf_results}
\alias{noconf_results}
\title{compiles the final results}
\usage{
noconf_results(r)
}
\description{
compiles the results into one table for all permutations of the simulation
}
