% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no-confounding-methods.R
\name{noconf_sample}
\alias{noconf_sample}
\title{perform sampling and coding of treatment for no confoudning policy simulations}
\usage{
noconf_sample(single_simulation)
}
\arguments{
\item{single_simulation}{object created from SimConfig$setup_single_simulation()}
}
\description{
todo
}
