% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optic_model_class.R
\name{optic_model}
\alias{optic_model}
\title{Optic Model}
\usage{
optic_model(name, type, call, formula, args, se_adjust)
}
\arguments{
\item{name}{describe}

\item{type}{describe}

\item{call}{describe}

\item{formula}{describe}

\item{args}{describe}

\item{se_adjust}{describe}
}
\value{
optic_model (optic_model) object
}
\description{
Optic Model
}
