% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optic-simulation.r
\docType{data}
\name{OpticConfig}
\alias{OpticConfig}
\title{Configuration object for OPTIC Simulation project}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
OpticConfig
}
\arguments{
\item{data}{data.frame or tibble object; must contain the following named 
fields: "population", "deaths", "year", "state"}

\item{model_type}{Currently supported: "linear", "log", or "loglinear"}

\item{method_call}{will be used with do.call to run your model with provided
parameters and additional arguments. Currently supported calls: "lm", "glm",
"glm.nb", "geeglm"}

\item{model_params}{list of named arguments and values to be passed to the
method_call specified.}

\item{outcome}{charcter name of field on data that should be used as outcome}

\item{iters}{number of iterations to run model}

\item{target_deaths}{number of deaths to use when calculating effect magnitude}

\item{effect_direction}{one of "null", "pos", "neg"}

\item{n_states}{number of states to sample as treatment; randomly sampled each iteration}

\item{policy_speed}{"instant" or "slow"}

\item{number_implementation_years}{required if policy_speed is slow, number of years
until policy is fully phased in}

\item{set_seed}{if TRUE, will set and store a seed prior to sampling and running
simulation for future replication}

\item{time_periods}{limit the years available for sampling by specifying a range,
e.g.: 2003:2013}
}
\description{
Configuration object for OPTIC Simulation project
}
\keyword{datasets}
