% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval-flag.r
\name{pval_flag}
\alias{pval_flag}
\title{create binary indicator for significance of p-value based on given level}
\usage{
pval_flag(p, level = 0.05)
}
\arguments{
\item{p}{vector of p-values}

\item{level}{threshold for significance; default is 0.05}
}
\description{
create binary indicator for significance of p-value based on given level
}
