% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-simulation.r
\name{run_simulation}
\alias{run_simulation}
\title{TODO: docstring}
\usage{
run_simulation(data, model_type, target_deaths, iters, n_states,
  effect_direction, policy_speed)
}
\arguments{
\item{data}{data.frame or tibble object containing required fields (see Details for more)}

\item{model_type}{"linear" (default), "log", or "log-lin"}

\item{target_deaths}{target number of change in deaths per 100,000 to use when calculating true effect}

\item{iters}{number of iterations to run}

\item{n_states}{number of states to treat}

\item{policy_speed}{"instant" (default) or "slow"}
}
\description{
TODO: docstring
}
