% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection-bias-methods.R
\name{selbias_results}
\alias{selbias_results}
\title{compiles the final results}
\usage{
selbias_results(r)
}
\description{
compiles the results into one table for all permutations of the simulation
}
