% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection-bias-methods.R
\name{selbias_sample}
\alias{selbias_sample}
\title{TODO: documentaion for selection bias sampling method}
\usage{
selbias_sample(single_simulation)
}
\description{
TODO: documentaion for selection bias sampling method
}
