% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-config.R
\name{SimConfig}
\alias{SimConfig}
\title{Configuration object for OPTIC Simulation project}
\description{
Configuration object for OPTIC Simulation project

Configuration object for OPTIC Simulation project
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SimConfig$new()}}
\item \href{#method-setup_single_simulation}{\code{SimConfig$setup_single_simulation()}}
\item \href{#method-print}{\code{SimConfig$print()}}
\item \href{#method-clone}{\code{SimConfig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimConfig$new(
  data,
  unit_var,
  time_var,
  model_call,
  model_args,
  model_formula,
  iters,
  effect_direction,
  effect_magnitude,
  n_units,
  policy_speed,
  se_adjust,
  concurrent,
  lag_outcome,
  n_implementation_periods,
  time_period_restriction,
  correction_factors,
  rhos,
  change_code_treatment
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setup_single_simulation"></a>}}
\subsection{Method \code{setup_single_simulation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimConfig$setup_single_simulation(combination)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimConfig$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
