% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch-simulations.R
\name{simulate.OpticSim}
\alias{simulate.OpticSim}
\title{Execute simulations defined in a optic_simulation object}
\usage{
\method{simulate}{OpticSim}(
  object,
  use_future = FALSE,
  seed = NULL,
  failure = NULL,
  verbose = 0,
  ...
)
}
\arguments{
\item{object}{Simulation scenarios created using optic_simulation}

\item{use_future}{Runs simulation scenarios in parallel. Default FALSE, set to TRUE if you have already setup a future
plan (e.g., multiprocess, cluster, etc) and would like for the iterations to
be run in parallel.}

\item{seed}{Specified as either NULL or a numeric. Sets a seed, which is becomes an index in results, for
each indepdendent set of simulations in optic_simulation.}

\item{verbose}{Default TRUE. IF TRUE, provides details on what's currently running.}
}
\description{
Execute simulations defined in a optic_simulation object
}
