% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch-simulations.R
\name{simulate.SimConfig}
\alias{simulate.SimConfig}
\title{Execute simulations defined in a SimConfig object}
\usage{
\method{simulate}{SimConfig}(
  sim_config,
  use_future = FALSE,
  seed = NULL,
  failure = NULL,
  verbose = 0,
  ...
)
}
\arguments{
\item{sim_config}{SimConfig object created using 'create_config'}

\item{use_future}{default FALSE, set to TRUE if you have already setup a future
plan (e.g., multiprocess, cluster, etc) and would like for the iterations to
be run in parallel}

\item{seed}{if not null, will use the index of the simulation * seed provided for
each indepdendent set of simulations}

\item{verbose}{default TRUE, have the dispatcher tell you what's currently running}
}
\description{
Execute simulations defined in a SimConfig object
}
