% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-s-error.r
\name{type_s_error}
\alias{type_s_error}
\title{Calculate type S error - how often model gets direction of effect wrong}
\usage{
type_s_error(betas, pvals, effect_direction)
}
\arguments{
\item{betas}{vector of regression coefficients}

\item{pvals}{vector of p values for betas}

\item{effect_direction}{direction of true effect}
}
\description{
number of significant betas of certain direction divided by all significant betas
}
