% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-treatment-effect.r
\name{apply_treatment_effect}
\alias{apply_treatment_effect}
\title{apply specified treatment effect and transform outcome to match model type}
\usage{
apply_treatment_effect(
  x,
  model_call,
  model_formula,
  te,
  effect_direction,
  concurrent
)
}
\arguments{
\item{te}{output of effect_magnitude, the true effect}

\item{ConfigObject}{R6Class object of class "OpticConfig"}
}
\description{
Calls appropriate method depending on model type (class) of OpticConfig object
}
