% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concurrent-methods.R
\name{concurrent_sample}
\alias{concurrent_sample}
\title{perform sampling and coding of treatment for concurrent policy simulations}
\usage{
concurrent_sample(single_simulation)
}
\arguments{
\item{single_simulation}{object created from SimConfig$setup_single_simulation()}
}
\description{
samples the treated units (e.g., states) randomly and then generates enacted dates for two policies with a prespeficied average mean difference in length between enactment dates of the two; also computes the needed levels and change levels coding for the time-varying vectors of the treatment/exposure variables representing those two policies
}
