% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.r
\name{coverage}
\alias{coverage}
\title{calculate percentage of time the 95 percent CI actually covers the true
treatment effect estimate}
\usage{
coverage(beta, se, cf, te)
}
\arguments{
\item{beta}{vector of coefficients}

\item{se}{vector of standard errors for coefficients provided in beta}

\item{cf}{correction factor (single constant)}

\item{te}{true treatment effect}
}
\description{
calculate percentage of time the 95 percent CI actually covers the true
treatment effect estimate
}
