% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-distance-avgs.R
\name{get_distance_avgs}
\alias{get_distance_avgs}
\title{get distance between concurrent policies for each treated unit and summarize}
\usage{
get_distance_avgs(treated_units, units = "days")
}
\arguments{
\item{treated_units}{list containing the sampled treated units for a concurrent
simulation run}

\item{units}{time units for distance calculation, default is "days"}
}
\description{
get distance between concurrent policies for each treated unit and summarize
}
