% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iter-results.r
\name{iter_results}
\alias{iter_results}
\title{iter results}
\usage{
iter_results(m, ConfigObject)
}
\arguments{
\item{m}{model object}

\item{ConfigObject}{R6 class object for simulation config}
}
\description{
iter results
}
