% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iter-results-concurrent-wjointeff.R
\name{iter_results_concurrent_wjointeff}
\alias{iter_results_concurrent_wjointeff}
\title{iter results}
\usage{
iter_results_concurrent_wjointeff(model_simulation)
}
\arguments{
\item{model_simulation}{}

\item{m}{model object}
}
\description{
iter results
}
