% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optic-simulation-helper.R
\name{optic_simulation}
\alias{optic_simulation}
\title{Create a configuration object used to run simulations}
\usage{
optic_simulation(
  x,
  models,
  iters,
  unit_var,
  time_var,
  effect_magnitude,
  n_units,
  effect_direction,
  policy_speed,
  prior_control,
  treat_var,
  n_implementation_periods,
  rhos,
  years_apart,
  ordered,
  method,
  method_sample,
  method_model,
  method_results,
  method_pre_model,
  method_post_model,
  globals = NULL,
  verbose = TRUE
)
}
\arguments{
\item{x}{Empirical data used to simulate synthetic datasets with specified treatment effect.}

\item{models}{List of `optic_model` objects that should be run for each iteration and simulation scenario.
The elements must be created using the `optic_model` function.}

\item{iters}{Number of iterations for each simulation scenario.}

\item{unit_var}{char(1) variable specifying the unit to be simulate (i.e., "state")}

\item{time_var}{char(1) variable specifying the time variable (i.e., "year")}

\item{effect_magnitude}{document. list of scenarios to consider}

\item{n_units}{integer. document.}

\item{effect_direction}{character vector containing the effects direction to consider. can include "null", "pos" and/or "neg"}

\item{policy_speed}{document. either "instant" or "slow"}

\item{prior_control}{character(n) document. apparently only used for no_confounding}

\item{treat_var}{character(1). document. apparently only used for no_confounding.}

\item{n_implementation_periods}{document}

\item{rhos}{document}

\item{years_apart}{document}

\item{ordered}{boolean. document}

\item{method}{character. either no_confounding, confounding or concurrent}

\item{method_sample}{function for sampling treated units, should modify the
single_simulation$data object}

\item{method_model}{function to run model, default performs a do.call on the 
model_call, passing model_formula and model_args if provided. Resulting 
model object is added to single_simulation list under named element
"model_result"}

\item{method_results}{function that takes the single_simulation object and
the return value is what it returned from run_iteration}

\item{method_pre_model}{optional function that will be applied single_simulation
object prior to modeling}

\item{method_post_model}{optional function for any post-processing on 
single_simulation object}

\item{globals}{Additional globals to pass to the simulate function, such as parallelization packages or additional R packages used by method calls.}

\item{verbose}{Default True. If TRUE, provides summary details on simulation runs across iterations}
}
\description{
Performs validation on inputs and produces a configuration object
    that contains all required parameters to dispatch simulation runs for the empirical data
    provided.
}
\details{
The resulting configuration object is used to pass simulation scenarios to the 'simulate' function. Provided as a convenience function to the user
         so they can investigate simulation arguments prior to running models.
}
