% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optic-simulation.r
\name{OpticConfig}
\alias{OpticConfig}
\title{Configuration object for OPTIC Simulation project}
\description{
Configuration object for OPTIC Simulation project

Configuration object for OPTIC Simulation project
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{TODO}

\item{\code{model_type}}{TODO}

\item{\code{method_call}}{TODO}

\item{\code{model_params}}{TODO}

\item{\code{outcome}}{TODO}

\item{\code{iters}}{TODO}

\item{\code{target_deaths}}{TODO}

\item{\code{effect_direction}}{TODO}

\item{\code{n_states}}{TODO}

\item{\code{policy_speed}}{TODO}

\item{\code{total_population}}{TODO}

\item{\code{total_deaths}}{TODO}

\item{\code{number_implementation_years}}{TODO}

\item{\code{seed}}{TODO}

\item{\code{time_periods}}{TODO}

\item{\code{iter_results}}{TODO}

\item{\code{summary_results}}{TODO}

\item{\code{lag_outcome}}{TODO}

\item{\code{se_adjust}}{TODO}

\item{\code{concurrent}}{TODO}

\item{\code{correction_factors}}{TODO}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OpticConfig$new()}}
\item \href{#method-print}{\code{OpticConfig$print()}}
\item \href{#method-clone}{\code{OpticConfig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpticConfig$new(
  data,
  model_type,
  method_call,
  model_params,
  iters,
  target_deaths,
  effect_direction,
  n_states,
  policy_speed,
  outcome,
  se_adjust = c("huber", "cluster", "huber-cluster"),
  concurrent = FALSE,
  lag_outcome = FALSE,
  number_implementation_years = NULL,
  set_seed = FALSE,
  time_periods = NULL,
  correction_factors = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{TODO}

\item{\code{model_type}}{Currently supported: "linear", "log", or "loglinear"}

\item{\code{method_call}}{will be used with do.call to run your model with provided
parameters and additional arguments. Currently supported calls: "lm", "glm",
"glm.nb", "geeglm"}

\item{\code{model_params}}{list of named arguments and values to be passed to the
method_call specified.}

\item{\code{iters}}{number of iterations to run model}

\item{\code{target_deaths}}{number of deaths to use when calculating effect magnitude}

\item{\code{effect_direction}}{one of "null", "pos", "neg"}

\item{\code{n_states}}{number of states to sample as treatment; randomly sampled each iteration}

\item{\code{policy_speed}}{"instant" or "slow"}

\item{\code{outcome}}{charcter name of field on data that should be used as outcome}

\item{\code{se_adjust}}{TODO}

\item{\code{concurrent}}{TODO
Print method of optic R6 config object}

\item{\code{lag_outcome}}{TODO}

\item{\code{number_implementation_years}}{required if policy_speed is slow, number of years
until policy is fully phased in}

\item{\code{set_seed}}{if TRUE, will set and store a seed prior to sampling and running
simulation for future replication}

\item{\code{time_periods}}{limit the years available for sampling by specifying a range,
e.g.: 2003:2013}

\item{\code{correction_factors}}{TODO}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpticConfig$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{TODO}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpticConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
