% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optic-simulation.r
\docType{data}
\name{OpticConfig}
\alias{OpticConfig}
\title{Configuration object for OPTIC Simulation project}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
OpticConfig
}
\arguments{
\item{model_type}{Currently supported: "linear", "log", or "loglinear"}

\item{method_call}{will be used with do.call to run your model with provided
parameters and additional arguments. Currently supported calls: "lm", "glm",
"glm.nb", "geeglm"}

\item{model_params}{list of named arguments and values to be passed to the
method_call specified.}

\item{outcome}{charcter name of field on data that should be used as outcome}

\item{iters}{number of iterations to run model}

\item{target_deaths}{number of deaths to use when calculating effect magnitude}

\item{effect_direction}{one of "null", "pos", "neg"}

\item{n_states}{number of states to sample as treatment; randomly sampled each iteration}

\item{policy_speed}{"instant" or "slow"}

\item{number_implementation_years}{required if policy_speed is slow, number of years
until policy is fully phased in}

\item{set_seed}{if TRUE, will set and store a seed prior to sampling and running
simulation for future replication}

\item{time_periods}{limit the years available for sampling by specifying a range,
e.g.: 2003:2013}

\item{correction_factors}{TODO}

\item{lag_outcome}{TODO}

\item{data}{TODO}

\item{se_adjust}{TODO}

\item{concurrent}{TODO
Print method of optic R6 config object}

\item{...}{TODO}
}
\description{
Configuration object for OPTIC Simulation project
}
\section{Fields}{

\describe{
\item{\code{data}}{TODO}

\item{\code{model_type}}{TODO}

\item{\code{method_call}}{TODO}

\item{\code{model_params}}{TODO}

\item{\code{outcome}}{TODO}

\item{\code{iters}}{TODO}

\item{\code{target_deaths}}{TODO}

\item{\code{effect_direction}}{TODO}

\item{\code{n_states}}{TODO}

\item{\code{policy_speed}}{TODO}

\item{\code{total_population}}{TODO}

\item{\code{total_deaths}}{TODO}

\item{\code{number_implementation_years}}{TODO}

\item{\code{seed}}{TODO}

\item{\code{time_periods}}{TODO}

\item{\code{iter_results}}{TODO}

\item{\code{summary_results}}{TODO}

\item{\code{lag_outcome}}{TODO}

\item{\code{se_adjust}}{TODO}

\item{\code{concurrent}}{TODO}

\item{\code{correction_factors}}{TODO}
}}

\keyword{datasets}
