% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-config.R
\name{SimConfig}
\alias{SimConfig}
\title{Configuration object for OPTIC Simulation project}
\description{
Configuration object for OPTIC Simulation project

Configuration object for OPTIC Simulation project
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SimConfig$new()}}
\item \href{#method-setup_single_simulation}{\code{SimConfig$setup_single_simulation()}}
\item \href{#method-print}{\code{SimConfig$print()}}
\item \href{#method-clone}{\code{SimConfig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimConfig$new(
  data,
  models,
  method_class,
  method_sample,
  method_te,
  method_pre_model,
  method_model,
  method_post_model,
  method_results,
  iters,
  params
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setup_single_simulation"></a>}}
\if{latex}{\out{\hypertarget{method-setup_single_simulation}{}}}
\subsection{Method \code{setup_single_simulation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimConfig$setup_single_simulation(i)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimConfig$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
